contoso-call-center-backend/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Devin AI <158243242+devin-ai-integration[bot]@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.14"}
psycopg = {extras = ["binary"], version = "^3.2.9"}
faker = "^37.4.0"
pydub = "^0.25.1"
numpy = "^2.3.1"
scipy = "^1.16.0"
python-multipart = "^0.0.20"
azure-cognitiveservices-speech = "^1.44.0"
python-dotenv = "^1.1.1"
gender-guesser = "^0.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
